pyproject.toml
[project]
name = "splleed"
version = "0.1.0"
description = "LLM inference benchmarking harness with pluggable backends"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "splleed contributors" }]
keywords = ["llm", "benchmark", "inference", "vllm", "performance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "httpx>=0.25",
    "rich>=13.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
vllm = ["vllm"]
tgi = ["text-generation"]
transformers = ["transformers", "torch"]
cloud = ["skypilot"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
splleed = "splleed.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/splleed"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.407",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "vulture>=2.14",
]
